pyproject.toml
[tool.poetry]
name = "ice-breaker"
version = "0.1.0"
description = ""
authors = ["Mario Toribio <yuniormtr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.13"
ruff = "^0.3.4"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.1"
langchain-google-genai = "^1.0.1"
pip = "^24.0"
install = "^1.3.5"
beautifulsoup4 = "^4.12.3"
black = "^24.3.0"
langchainhub = "^0.1.15"
google-search-results = "^2.4.2"
tweepy = "^4.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
